+++ src/app/(auth)/_components/AuthForm.tsx
@@ -5,8 +5,7 @@ import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Button, Text } from "@radix-ui/themes";
 import Link from "next/link";
-import { useState } from "react";
-import { HiEye, HiEyeOff } from "react-icons/hi";
+import PasswordInput from "./PasswordInput";
 
 /**
  * 認証フォームのバリデーションスキーマ
@@ -57,23 +56,6 @@ export default function AuthForm({ formType, onSubmit }: AuthFormProps) {
     mode: "onTouched",
   });
 
-  const [showPassword, setShowPassword] = useState(false);
-  const password = watch("password", "");
-
-  /**
-   * パスワードの要件チェック
-   * - 8文字以上
-   * - アルファベットを含む
-   * - 数字を含む
-   * - 特殊文字を含む
-   */
-  const passwordRequirements = {
-    length: password.length >= 8,
-    letter: /[A-Za-z]/.test(password),
-    number: /\d/.test(password),
-    special: /[@$!%*#?&]/.test(password),
-  };
-
   /** フォーム送信処理 */
   const onSubmitHandler = async (data: FormData) => {
     try {
@@ -82,11 +64,6 @@ export default function AuthForm({ formType, onSubmit }: AuthFormProps) {
     } catch (error) {
       console.error(error);
     }
-  };
-
-  /** パスワードの表示/非表示切り替え */
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
   };
 
   return (
@@ -114,76 +91,14 @@ export default function AuthForm({ formType, onSubmit }: AuthFormProps) {
             <p className="mt-1 text-sm text-red-600">{errors.email.message}</p>
           )}
         </div>
-        <div>
-          <label
-            htmlFor="password"
-            className="mb-2 block text-sm font-medium text-gray-900"
-          >
-            パスワード
-          </label>
-          <div className="relative">
-            <input
-              type={showPassword ? "text" : "password"}
-              id="password"
-              placeholder="••••••••"
-              className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-blue-500 focus:ring-blue-500 disabled:cursor-not-allowed disabled:opacity-50"
-              disabled={isSubmitting}
-              {...register("password")}
-            />
-            <button
-              type="button"
-              onClick={togglePasswordVisibility}
-              className="absolute inset-y-0 right-0 flex cursor-pointer items-center px-3 text-gray-600"
-              disabled={isSubmitting}
-            >
-              {showPassword ? (
-                <HiEyeOff className="h-5 w-5" />
-              ) : (
-                <HiEye className="h-5 w-5" />
-              )}
-            </button>
-          </div>
-          {errors.password && (
-            <p className="mt-1 text-sm text-red-600">
-              {errors.password.message}
-            </p>
-          )}
-          {formType === "signup" && (
-            <div className="mt-2 space-y-1 text-sm text-gray-600">
-              <p>パスワードは以下の要件を満たす必要があります：</p>
-              <ul className="list-inside space-y-1">
-                <li
-                  className={
-                    passwordRequirements.length ? "text-green-600" : ""
-                  }
-                >
-                  ✓ 8文字以上
-                </li>
-                <li
-                  className={
-                    passwordRequirements.letter ? "text-green-600" : ""
-                  }
-                >
-                  ✓ アルファベットを含む
-                </li>
-                <li
-                  className={
-                    passwordRequirements.number ? "text-green-600" : ""
-                  }
-                >
-                  ✓ 数字を含む
-                </li>
-                <li
-                  className={
-                    passwordRequirements.special ? "text-green-600" : ""
-                  }
-                >
-                  ✓ 記号（@$!%*#?&）を含む
-                </li>
-              </ul>
-            </div>
-          )}
-        </div>
+
+        <PasswordInput
+          register={register}
+          error={errors.password}
+          isSubmitting={isSubmitting}
+          showRequirements={formType === "signup"}
+          watch={watch}
+        />
 
         <div>
           <Button
